+++ klipper/motion_wizard.py
@@ -7,7 +7,7 @@
 import time
 
 
-PLUGIN_VERSION = "0.2.0"
+PLUGIN_VERSION = "0.2.1"
 
 
 class MotionWizard: